// src/store/mutation-types.ts
export enum MutationTypes {
  // Root
  SET_START_DATE = "SET_START_DATE",
  SET_END_DATE = "SET_END_DATE",
  SET_ACTIVE_TIME_ENTRY = "SET_ACTIVE_TIME_ENTRY",
  SET_SELECTED_TIME_ENTRY = "SET_SELECTED_TIME_ENTRY",
  SET_TIME_ENTRIES = "SET_TIME_ENTRIES",
  SET_STATISTICS = "SET_STATISTICS",

  // Auth
  SET_USER = "SET_USER",
  SET_TOKEN = "SET_TOKEN",
  SET_CURRENT_TEAM_ID = "SET_CURRENT_TEAM_ID",
}